requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
pillow>=9.0.0
pydantic>=2.0.0
tenacity>=8.0.0
yt-dlp>=2023.0.0
